refactor: change menu and route setup structure based on config
